Project transfer fix for ES indexing

See merge request https://gitlab.com/gitlab-org/gitlab/-/merge_requests/152650

Merged-by: Dmitry Gruzd <dgruzd@gitlab.com>
Approved-by: John Mason <9717668-johnmason@users.noreply.gitlab.com>
Approved-by: Dmitry Gruzd <dgruzd@gitlab.com>

+++ ee/app/workers/elastic_delete_project_worker.rb
@@ -15,6 +15,8 @@ class ElasticDeleteProjectWorker
   def perform(project_id, es_id, options = {})
     options = options.with_indifferent_access
     remove_project_document(project_id, es_id, options) if options.fetch(:delete_project, true)
+    return if options.fetch(:project_only, false)
+
     remove_children_documents(project_id, es_id, options)
     helper.remove_wikis_from_the_standalone_index(project_id, 'Project', options[:namespace_routing_id]) # Wikis have different routing that's why one more query is needed.
     IndexStatus.for_project(project_id).delete_all
